@@ -6,69 +6,58 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace CamoDotNet.Tests
+namespace CamoDotNet.Tests;
+
+public class CamoServerSignatureFacts
+    : CamoServerFactsBase
 {
-    public class CamoServerSignatureFacts
-        : CamoServerFactsBase
+    [Fact]
+    public async Task Returns404NotFoundForChecksumMismatchWithPathFormat()
     {
-        [Fact]
-        public async Task Returns404NotFoundForChecksumMismatchWithPathFormat()
-        {
-            using (var server = CreateServer())
-            {
-                var response = await server.CreateClient().GetAsync("/74657374/68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f4e754765742f486f6d652f6465762f7265736f75726365732f6e756765742e706e67");
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync("/74657374/68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f4e754765742f486f6d652f6465762f7265736f75726365732f6e756765742e706e67");
 
-                Assert.Contains("checksum mismatch", await response.Content.ReadAsStringAsync());
-                Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-            }
-        }
-
-        [Fact]
-        public async Task Returns404NotFoundForChecksumMismatchWithQueryFormat()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().GetAsync("/74657374?url=http%3A%2F%2Fwww.nuget.org%2Ftest");
-
-                Assert.Contains("checksum mismatch", await response.Content.ReadAsStringAsync());
-                Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-            }
-        }
+        Assert.Contains("checksum mismatch", await response.Content.ReadAsStringAsync());
+        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
     }
-    public class CamoServerDefaultDocumentFacts
-        : CamoServerFactsBase
+
+    [Fact]
+    public async Task Returns404NotFoundForChecksumMismatchWithQueryFormat()
     {
-        [Fact]
-        public async Task Returns405MethodNotAllowedForPostToRootUrl()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().PostAsync("/", new StringContent("data"));
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync("/74657374?url=http%3A%2F%2Fwww.nuget.org%2Ftest");
 
-                Assert.Equal(HttpStatusCode.MethodNotAllowed, response.StatusCode);
-            }
-        }
+        Assert.Contains("checksum mismatch", await response.Content.ReadAsStringAsync());
+        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
+    }
+}
+public class CamoServerDefaultDocumentFacts
+    : CamoServerFactsBase
+{
+    [Fact]
+    public async Task Returns405MethodNotAllowedForPostToRootUrl()
+    {
+        using var server = CreateServer();
+        var response = await server.CreateClient().PostAsync("/", new StringContent("data"));
 
-        [Fact]
-        public async Task Returns200OkForRootUrl()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().GetAsync("/");
+        Assert.Equal(HttpStatusCode.MethodNotAllowed, response.StatusCode);
+    }
 
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            }
-        }
+    [Fact]
+    public async Task Returns200OkForRootUrl()
+    {
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync("/");
 
-        [Fact]
-        public async Task Returns200OkForFavIconUrl()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().GetAsync("/favicon.ico");
+        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+    }
 
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            }
-        }
+    [Fact]
+    public async Task Returns200OkForFavIconUrl()
+    {
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync("/favicon.ico");
+
+        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
     }
 }
