@@ -6,44 +6,37 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace CamoDotNet.Tests
+namespace CamoDotNet.Tests;
+
+public class CamoServerProxyFacts
+    : CamoServerFactsBase
 {
-    public class CamoServerProxyFacts
-        : CamoServerFactsBase
+    [Fact]
+    public async Task ProxiesValidUrl()
     {
-        [Fact]
-        public async Task ProxiesValidUrl()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().GetAsync(GenerateSignedUrl("https://raw.githubusercontent.com/NuGet/Home/dev/meta/resources/nuget.png"));
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync(GenerateSignedUrl("https://raw.githubusercontent.com/NuGet/Home/dev/meta/resources/nuget.png"));
 
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                Assert.IsType<StreamContent>(response.Content);
-            }
-        }
+        Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        Assert.IsType<StreamContent>(response.Content);
+    }
 
-        [Fact]
-        public async Task DoesNotProxyInvalidContentType()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().GetAsync(GenerateSignedUrl("https://www.github.com"));
+    [Fact]
+    public async Task DoesNotProxyInvalidContentType()
+    {
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync(GenerateSignedUrl("https://www.github.com"));
 
-                Assert.Contains("Non-Image content-type returned", await response.Content.ReadAsStringAsync());
-                Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-            }
-        }
+        Assert.Contains("Non-Image content-type returned", await response.Content.ReadAsStringAsync());
+        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
+    }
 
-        [Fact]
-        public async Task DoesNotProxyInvalidUrl()
-        {
-            using (var server = CreateServer())
-            {
-                HttpResponseMessage response = await server.CreateClient().GetAsync(GenerateSignedUrl("https://www.github.com.thisdoesnotexist"));
+    [Fact]
+    public async Task DoesNotProxyInvalidUrl()
+    {
+        using var server = CreateServer();
+        var response = await server.CreateClient().GetAsync(GenerateSignedUrl("https://www.github.com.thisdoesnotexist"));
                 
-                Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
-            }
-        }
+        Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
     }
 }
